roachprod: separate certificate distribution from start
