+++ pkg/cmd/roachprod/install/cockroach.go
@@ -11,9 +11,7 @@
 package install
 
 import (
-	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -125,140 +123,7 @@ func (r Cockroach) Start(c *SyncedCluster, extraArgs []string) {
 	}
 
 	if c.Secure && bootstrapped {
-		dir := ""
-		if c.IsLocal() {
-			dir = `${HOME}/local/1`
-		}
-
-		// Check to see if the certs have already been initialized.
-		var existsErr error
-		display := fmt.Sprintf("%s: checking certs", c.Name)
-		c.Parallel(display, 1, 0, func(i int) ([]byte, error) {
-			sess, err := c.newSession(1)
-			if err != nil {
-				return nil, err
-			}
-			defer sess.Close()
-			_, existsErr = sess.CombinedOutput(`test -e ` + filepath.Join(dir, `certs.tar`))
-			return nil, nil
-		})
-
-		if existsErr != nil {
-			// Gather the internal IP addresses for every node in the cluster, even
-			// if it won't be added to the cluster itself we still add the IP address
-			// to the node cert.
-			var msg string
-			display := fmt.Sprintf("%s: initializing certs", c.Name)
-			nodes := allNodes(len(c.VMs))
-			var ips []string
-			if !c.IsLocal() {
-				ips = make([]string, len(nodes))
-				c.Parallel("", len(nodes), 0, func(i int) ([]byte, error) {
-					var err error
-					ips[i], err = c.GetInternalIP(nodes[i])
-					return nil, errors.Wrapf(err, "IPs")
-				})
-			}
-
-			// Generate the ca, client and node certificates on the first node.
-			c.Parallel(display, 1, 0, func(i int) ([]byte, error) {
-				sess, err := c.newSession(1)
-				if err != nil {
-					return nil, err
-				}
-				defer sess.Close()
-
-				var nodeNames []string
-				if c.IsLocal() {
-					// For local clusters, we only need to add one of the VM IP addresses.
-					nodeNames = append(nodeNames, "$(hostname)", c.VMs[0])
-				} else {
-					// Add both the local and external IP addresses, as well as the
-					// hostnames to the node certificate.
-					nodeNames = append(nodeNames, ips...)
-					nodeNames = append(nodeNames, c.VMs...)
-					for i := range c.VMs {
-						nodeNames = append(nodeNames, fmt.Sprintf("%s-%04d", c.Name, i+1))
-					}
-				}
-
-				var cmd string
-				if c.IsLocal() {
-					cmd = `cd ${HOME}/local/1 ; `
-				}
-				cmd += fmt.Sprintf(`
-rm -fr certs
-mkdir -p certs
-%[1]s cert create-ca --certs-dir=certs --ca-key=certs/ca.key
-%[1]s cert create-client root --certs-dir=certs --ca-key=certs/ca.key
-%[1]s cert create-node localhost %[2]s --certs-dir=certs --ca-key=certs/ca.key
-tar cvf certs.tar certs
-`, cockroachNodeBinary(c, 1), strings.Join(nodeNames, " "))
-				if out, err := sess.CombinedOutput(cmd); err != nil {
-					msg = fmt.Sprintf("%s: %v", out, err)
-				}
-				return nil, nil
-			})
-
-			if msg != "" {
-				fmt.Fprintln(os.Stderr, msg)
-				os.Exit(1)
-			}
-
-			var tmpfileName string
-			if c.IsLocal() {
-				tmpfileName = os.ExpandEnv(filepath.Join(dir, "certs.tar"))
-			} else {
-				// Retrieve the certs.tar that was created on the first node.
-				tmpfile, err := ioutil.TempFile("", "certs")
-				if err != nil {
-					fmt.Fprintln(os.Stderr, err)
-					os.Exit(1)
-				}
-				_ = tmpfile.Close()
-				defer func() {
-					_ = os.Remove(tmpfile.Name()) // clean up
-				}()
-
-				if err := func() error {
-					return c.scp(fmt.Sprintf("%s@%s:certs.tar", c.user(1), c.host(1)), tmpfile.Name())
-				}(); err != nil {
-					fmt.Fprintln(os.Stderr, err)
-					os.Exit(1)
-				}
-
-				tmpfileName = tmpfile.Name()
-			}
-
-			// Read the certs.tar file we just downloaded. We'll be piping it to the
-			// other nodes in the cluster.
-			certsTar, err := ioutil.ReadFile(tmpfileName)
-			if err != nil {
-				fmt.Fprintln(os.Stderr, err)
-				os.Exit(1)
-			}
-
-			// Skip the the first node which is where we generated the certs.
-			nodes = nodes[1:]
-			c.Parallel(display, len(nodes), 0, func(i int) ([]byte, error) {
-				sess, err := c.newSession(nodes[i])
-				if err != nil {
-					return nil, err
-				}
-				defer sess.Close()
-
-				sess.SetStdin(bytes.NewReader(certsTar))
-				var cmd string
-				if c.IsLocal() {
-					cmd = fmt.Sprintf(`cd ${HOME}/local/%d ; `, nodes[i])
-				}
-				cmd += `tar xf -`
-				if out, err := sess.CombinedOutput(cmd); err != nil {
-					return nil, errors.Wrapf(err, "~ %s\n%s", cmd, out)
-				}
-				return nil, nil
-			})
-		}
+		c.DistributeCerts()
 	}
 
 	display := fmt.Sprintf("%s: starting", c.Name)
